requirements.txt
altgraph==0.17.5
hidapi==0.15.0
packaging==26.0
pefile==2024.8.26
pillow==12.2.0
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
pypiwin32==223
pystray==0.19.5
pywin32==311
pywin32-ctypes==0.2.3
six==1.17.0
winotify==1.1.0
setuptools<=81.0.0
